pyproject.toml
[tool.poetry]
name = "ye-meme-trader"
version = "0.1.0"
description = "Solana meme token trading bot focused on Ye-related tokens"
authors = ["Alessandro Artini"]

[tool.poetry.dependencies]
python = "^3.10"
solana = "0.30.1"
anchorpy = "0.18.0"
redis = "5.0.1"
python-telegram-bot = "20.8"
python-dotenv = "1.0.0"
aiohttp = "3.9.3"
asyncio = "3.4.3"
web3 = "6.15.1"

[tool.poetry.group.dev.dependencies]
pytest = "8.0.0"
pytest-asyncio = "0.23.5"
pytest-cov = "4.1.0"
black = "24.1.1"
isort = "5.13.2"
mypy = "1.8.0"
pylint = "3.0.3"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ye_meme_trader"
version = "0.1.0"
authors = [
  { name="Alessandro Artini", email="ale7998@gmail.com" },
]
description = "Ye Meme Token Trading Bot"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot>=20.0",
    "redis>=4.0.0",
    "requests>=2.25.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=0.19.0",
    "solana==0.30.2",
    "anchorpy==0.18.0",
    "asyncio>=3.4.3",
    "websockets>=10.0",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "sentry-sdk>=1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
